+++ internal/database/ent/db/applicationcommand/where.go
@@ -63,12 +63,12 @@ func IDContainsFold(id string) predicate.ApplicationCommand {
 }
 
 // CreatedAt applies equality check predicate on the "created_at" field. It's identical to CreatedAtEQ.
-func CreatedAt(v int) predicate.ApplicationCommand {
+func CreatedAt(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldEQ(FieldCreatedAt, v))
 }
 
 // UpdatedAt applies equality check predicate on the "updated_at" field. It's identical to UpdatedAtEQ.
-func UpdatedAt(v int) predicate.ApplicationCommand {
+func UpdatedAt(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldEQ(FieldUpdatedAt, v))
 }
 
@@ -88,82 +88,82 @@ func OptionsHash(v string) predicate.ApplicationCommand {
 }
 
 // CreatedAtEQ applies the EQ predicate on the "created_at" field.
-func CreatedAtEQ(v int) predicate.ApplicationCommand {
+func CreatedAtEQ(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldEQ(FieldCreatedAt, v))
 }
 
 // CreatedAtNEQ applies the NEQ predicate on the "created_at" field.
-func CreatedAtNEQ(v int) predicate.ApplicationCommand {
+func CreatedAtNEQ(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldNEQ(FieldCreatedAt, v))
 }
 
 // CreatedAtIn applies the In predicate on the "created_at" field.
-func CreatedAtIn(vs ...int) predicate.ApplicationCommand {
+func CreatedAtIn(vs ...int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldIn(FieldCreatedAt, vs...))
 }
 
 // CreatedAtNotIn applies the NotIn predicate on the "created_at" field.
-func CreatedAtNotIn(vs ...int) predicate.ApplicationCommand {
+func CreatedAtNotIn(vs ...int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldNotIn(FieldCreatedAt, vs...))
 }
 
 // CreatedAtGT applies the GT predicate on the "created_at" field.
-func CreatedAtGT(v int) predicate.ApplicationCommand {
+func CreatedAtGT(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldGT(FieldCreatedAt, v))
 }
 
 // CreatedAtGTE applies the GTE predicate on the "created_at" field.
-func CreatedAtGTE(v int) predicate.ApplicationCommand {
+func CreatedAtGTE(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldGTE(FieldCreatedAt, v))
 }
 
 // CreatedAtLT applies the LT predicate on the "created_at" field.
-func CreatedAtLT(v int) predicate.ApplicationCommand {
+func CreatedAtLT(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldLT(FieldCreatedAt, v))
 }
 
 // CreatedAtLTE applies the LTE predicate on the "created_at" field.
-func CreatedAtLTE(v int) predicate.ApplicationCommand {
+func CreatedAtLTE(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldLTE(FieldCreatedAt, v))
 }
 
 // UpdatedAtEQ applies the EQ predicate on the "updated_at" field.
-func UpdatedAtEQ(v int) predicate.ApplicationCommand {
+func UpdatedAtEQ(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldEQ(FieldUpdatedAt, v))
 }
 
 // UpdatedAtNEQ applies the NEQ predicate on the "updated_at" field.
-func UpdatedAtNEQ(v int) predicate.ApplicationCommand {
+func UpdatedAtNEQ(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldNEQ(FieldUpdatedAt, v))
 }
 
 // UpdatedAtIn applies the In predicate on the "updated_at" field.
-func UpdatedAtIn(vs ...int) predicate.ApplicationCommand {
+func UpdatedAtIn(vs ...int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldIn(FieldUpdatedAt, vs...))
 }
 
 // UpdatedAtNotIn applies the NotIn predicate on the "updated_at" field.
-func UpdatedAtNotIn(vs ...int) predicate.ApplicationCommand {
+func UpdatedAtNotIn(vs ...int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldNotIn(FieldUpdatedAt, vs...))
 }
 
 // UpdatedAtGT applies the GT predicate on the "updated_at" field.
-func UpdatedAtGT(v int) predicate.ApplicationCommand {
+func UpdatedAtGT(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldGT(FieldUpdatedAt, v))
 }
 
 // UpdatedAtGTE applies the GTE predicate on the "updated_at" field.
-func UpdatedAtGTE(v int) predicate.ApplicationCommand {
+func UpdatedAtGTE(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldGTE(FieldUpdatedAt, v))
 }
 
 // UpdatedAtLT applies the LT predicate on the "updated_at" field.
-func UpdatedAtLT(v int) predicate.ApplicationCommand {
+func UpdatedAtLT(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldLT(FieldUpdatedAt, v))
 }
 
 // UpdatedAtLTE applies the LTE predicate on the "updated_at" field.
-func UpdatedAtLTE(v int) predicate.ApplicationCommand {
+func UpdatedAtLTE(v int64) predicate.ApplicationCommand {
 	return predicate.ApplicationCommand(sql.FieldLTE(FieldUpdatedAt, v))
 }
