+++ internal/database/ent/db/userconnection/where.go
@@ -65,12 +65,12 @@ func IDContainsFold(id string) predicate.UserConnection {
 }
 
 // CreatedAt applies equality check predicate on the "created_at" field. It's identical to CreatedAtEQ.
-func CreatedAt(v int) predicate.UserConnection {
+func CreatedAt(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldEQ(FieldCreatedAt, v))
 }
 
 // UpdatedAt applies equality check predicate on the "updated_at" field. It's identical to UpdatedAtEQ.
-func UpdatedAt(v int) predicate.UserConnection {
+func UpdatedAt(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldEQ(FieldUpdatedAt, v))
 }
 
@@ -90,82 +90,82 @@ func Permissions(v string) predicate.UserConnection {
 }
 
 // CreatedAtEQ applies the EQ predicate on the "created_at" field.
-func CreatedAtEQ(v int) predicate.UserConnection {
+func CreatedAtEQ(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldEQ(FieldCreatedAt, v))
 }
 
 // CreatedAtNEQ applies the NEQ predicate on the "created_at" field.
-func CreatedAtNEQ(v int) predicate.UserConnection {
+func CreatedAtNEQ(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldNEQ(FieldCreatedAt, v))
 }
 
 // CreatedAtIn applies the In predicate on the "created_at" field.
-func CreatedAtIn(vs ...int) predicate.UserConnection {
+func CreatedAtIn(vs ...int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldIn(FieldCreatedAt, vs...))
 }
 
 // CreatedAtNotIn applies the NotIn predicate on the "created_at" field.
-func CreatedAtNotIn(vs ...int) predicate.UserConnection {
+func CreatedAtNotIn(vs ...int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldNotIn(FieldCreatedAt, vs...))
 }
 
 // CreatedAtGT applies the GT predicate on the "created_at" field.
-func CreatedAtGT(v int) predicate.UserConnection {
+func CreatedAtGT(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldGT(FieldCreatedAt, v))
 }
 
 // CreatedAtGTE applies the GTE predicate on the "created_at" field.
-func CreatedAtGTE(v int) predicate.UserConnection {
+func CreatedAtGTE(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldGTE(FieldCreatedAt, v))
 }
 
 // CreatedAtLT applies the LT predicate on the "created_at" field.
-func CreatedAtLT(v int) predicate.UserConnection {
+func CreatedAtLT(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldLT(FieldCreatedAt, v))
 }
 
 // CreatedAtLTE applies the LTE predicate on the "created_at" field.
-func CreatedAtLTE(v int) predicate.UserConnection {
+func CreatedAtLTE(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldLTE(FieldCreatedAt, v))
 }
 
 // UpdatedAtEQ applies the EQ predicate on the "updated_at" field.
-func UpdatedAtEQ(v int) predicate.UserConnection {
+func UpdatedAtEQ(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldEQ(FieldUpdatedAt, v))
 }
 
 // UpdatedAtNEQ applies the NEQ predicate on the "updated_at" field.
-func UpdatedAtNEQ(v int) predicate.UserConnection {
+func UpdatedAtNEQ(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldNEQ(FieldUpdatedAt, v))
 }
 
 // UpdatedAtIn applies the In predicate on the "updated_at" field.
-func UpdatedAtIn(vs ...int) predicate.UserConnection {
+func UpdatedAtIn(vs ...int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldIn(FieldUpdatedAt, vs...))
 }
 
 // UpdatedAtNotIn applies the NotIn predicate on the "updated_at" field.
-func UpdatedAtNotIn(vs ...int) predicate.UserConnection {
+func UpdatedAtNotIn(vs ...int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldNotIn(FieldUpdatedAt, vs...))
 }
 
 // UpdatedAtGT applies the GT predicate on the "updated_at" field.
-func UpdatedAtGT(v int) predicate.UserConnection {
+func UpdatedAtGT(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldGT(FieldUpdatedAt, v))
 }
 
 // UpdatedAtGTE applies the GTE predicate on the "updated_at" field.
-func UpdatedAtGTE(v int) predicate.UserConnection {
+func UpdatedAtGTE(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldGTE(FieldUpdatedAt, v))
 }
 
 // UpdatedAtLT applies the LT predicate on the "updated_at" field.
-func UpdatedAtLT(v int) predicate.UserConnection {
+func UpdatedAtLT(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldLT(FieldUpdatedAt, v))
 }
 
 // UpdatedAtLTE applies the LTE predicate on the "updated_at" field.
-func UpdatedAtLTE(v int) predicate.UserConnection {
+func UpdatedAtLTE(v int64) predicate.UserConnection {
 	return predicate.UserConnection(sql.FieldLTE(FieldUpdatedAt, v))
 }
