+++ internal/database/ent/schema/achievements_snapshot.go
@@ -23,7 +23,7 @@ func (AchievementsSnapshot) Fields() []ent.Field {
 		field.String("reference_id").NotEmpty(),
 		//
 		field.Int("battles"),
-		field.Int("last_battle_time"),
+		field.Int64("last_battle_time"),
 		field.JSON("data", types.AchievementsFrame{}),
 	)
 }
